refactor(analysis): move implementation of box size detection to new file box_config.rs
