+++ src/ans/mod.rs
@@ -1,4 +1,5 @@
 pub(crate) mod voronoy;
 pub(crate) mod analysis;
+pub(crate) mod box_config;
 mod libminio_rw;
 mod minio_input;
