+++ src/ans/voronoy.rs
@@ -1,17 +1,14 @@
-use std::{f32, io};
+use std::{f32};
 use std::fs::File;
 use std::io::{BufWriter, Write};
 use std::path::Path;
 use std::collections::BTreeMap;
-use xyzio::Reader;
 use std::error::Error;
-use crate::ans::analysis::BoxConfig;
-
 
 type Float = f32;
 type Inx = i32;
 
-const LATTICE_CONST: Float = 2.85532;
+pub const LATTICE_CONST: Float = 2.85532;
 
 /**
  * The coordinate offset of 8 minor lattices surrounding a major lattice.
@@ -63,68 +60,7 @@ const NORMAL_VECTOR: [(Float, Float, Float); 8] = [
 const D: Float = -3.0 / 4.0;
 
 
-// voronoy analysis method for BCC lattice and cube lattice.
-// todo: on data read error
-pub fn voronoy_ans<R: ?Sized>(input: &mut R, output: &str, on_data_loaded: impl Fn(usize),
-                              box_config: &BoxConfig, verbose: bool)
-    where R: io::Read
-{
-    let mut reader = Reader::new(input);
-    // todo read atom one by one and compute its index lattice.
-    let snapshot_result = reader.read_snapshot();
-
-    match snapshot_result {
-        Err(e) => {
-            println!("read input xyz file error: {:?}", e);
-        }
-        Ok(mut snapshot) => {
-            let atoms_size = snapshot.size();
-            on_data_loaded(atoms_size);
-            if atoms_size % 2 != 0 { // due to feature of BCC lattice
-                println!("bad atoms size");
-                return;
-            }
-
-            // determine box size
-            let (mut box_size_x, mut box_size_y, mut box_size_z) = (0, 0, 0);
-            if box_config.box_size.len() == 0 {
-                // automatically determine calculate box size via atoms position
-                let size_dim = cube_root(atoms_size / 2);
-                if size_dim != 0 { // it is a cube box
-                    box_size_x = size_dim;
-                    box_size_y = size_dim;
-                    box_size_z = size_dim;
-                } else {
-                    let (_size_x, _size_y, _size_z) = auto_get_box_size(&snapshot.atoms);
-                    box_size_x = _size_x;
-                    box_size_y = _size_y;
-                    box_size_z = _size_z;
-                }
-            } else {
-                box_size_x = box_config.box_size[0] as usize;
-                box_size_y = box_config.box_size[1] as usize;
-                box_size_z = box_config.box_size[2] as usize;
-            }
-
-            // check box size
-            if box_size_x == 0 || box_size_y == 0 || box_size_z == 0 {
-                println!("bad box size");
-                return;
-            }
-            if box_size_x * box_size_y * box_size_z != atoms_size {
-                println!("Warning: box size ({},{},{}) not match atoms size.", box_size_x, box_size_y, box_size_z);
-                return;
-            }
-            if verbose {
-                println!("box size: [{},{},{}]", box_size_x, box_size_y, box_size_z);
-            }
-            // do analysis
-            do_analysis_wrapper(output, box_size_x, box_size_y, box_size_z, &snapshot);
-        }
-    }
-}
-
-fn do_analysis_wrapper(output: &str, size_x: usize, size_y: usize, size_z: usize, mut snapshot: &xyzio::Snapshot) {
+pub fn do_analysis_wrapper(output: &str, size_x: usize, size_y: usize, size_z: usize, mut snapshot: &xyzio::Snapshot) {
     // prepare file
     let path = Path::new(output);
     let display = path.display();
@@ -137,67 +73,6 @@ fn do_analysis_wrapper(output: &str, size_x: usize, size_y: usize, size_z: usize
     let mut writer = BufWriter::new(file);
     do_analysis(&mut writer, size_x, size_y, size_z, &snapshot.atoms);
     writer.flush().unwrap();
-}
-
-// By passing the atoms position list,
-// then we can get the box size of simulation box, which can be used to calculating 1D lattice index.
-// If the box size in some dimension is not as desired, 0 will be return in the dimension.
-fn auto_get_box_size(atoms: &Vec<xyzio::Atom>) -> (usize, usize, usize) {
-    let mut x_min = f32::INFINITY; // todo can use f64 as float
-    let mut y_min = f32::INFINITY;
-    let mut z_min = f32::INFINITY;
-    let mut x_max = f32::NEG_INFINITY;
-    let mut y_max = f32::NEG_INFINITY;
-    let mut z_max = f32::NEG_INFINITY;
-    for i in 0..atoms.len() {
-        let x = atoms[i].x;
-        let y = atoms[i].y;
-        let z = atoms[i].z;
-
-        if x < x_min {
-            x_min = x;
-        }
-        if x > x_max {
-            x_max = x;
-        }
-
-        if y < y_min {
-            y_min = y;
-        }
-        if y > y_max {
-            y_max = y;
-        }
-
-        if z < z_min {
-            z_min = z;
-        }
-        if z > z_max {
-            z_max = z;
-        }
-    }
-
-    let min_index = voronoy(x_min, y_min, z_min);
-    let max_index = voronoy(x_max, y_max, z_max);
-    let (size_x_, size_y_, size_z_) = (max_index.0 - min_index.0, max_index.1 - min_index.1, max_index.2 - min_index.2);
-    let mut sizes = (0, 0, 0);
-
-    if size_x_ < 0 {
-        sizes.0 = 0;
-    } else {
-        sizes.0 = size_x_ as usize;
-    }
-    if size_y_ < 0 {
-        sizes.1 = 0;
-    } else {
-        sizes.1 = size_y_ as usize;
-    }
-
-    if size_z_ < 0 {
-        sizes.2 = 0;
-    } else {
-        sizes.2 = size_z_ as usize;
-    }
-    return sizes;
 }
 
 // in do_analysis, calculate atom's occupation by box size and its lattice index,
@@ -244,7 +119,7 @@ fn do_analysis(writer: &mut BufWriter<File>, box_x_size: usize, box_y_size: usiz
  * calculate coordinate of nearest lattice for a atom:
  * X,Y,Z is the position of atom.
  */
-fn voronoy(x: Float, y: Float, z: Float) -> (Inx, Inx, Inx) {
+pub fn voronoy(x: Float, y: Float, z: Float) -> (Inx, Inx, Inx) {
     let lat_coord_x = (x / LATTICE_CONST).round() as Inx;
     let mut lat_coord_y = (y / LATTICE_CONST).round() as Inx;
     let mut lat_coord_z = (z / LATTICE_CONST).round() as Inx;
@@ -270,51 +145,6 @@ fn voronoy(x: Float, y: Float, z: Float) -> (Inx, Inx, Inx) {
     return (lat_coord_x, lat_coord_y, lat_coord_z);
 }
 
-// return cube root of a positive integer number.
-fn cube_root(n: usize) -> usize {
-    let mut low = 1 as usize;
-    let mut high = n;
-    let mut mid = n;
-    while low <= high {
-        mid = (high + low) / 2;
-        if n < mid * mid * mid {
-            high = mid - 1;
-        } else if n == mid * mid * mid {
-            return mid;
-        } else {
-            low = mid + 1;
-        }
-    }
-    return 0;
-}
-
-#[cfg(test)]
-mod get_box_size_tests {
-    // Note this useful idiom: importing names from outer (for mod tests) scope.
-    use super::*;
-
-    #[test]
-    fn test_get_box_size_1() {
-        let mut atoms = Vec::new();
-        atoms.push(xyzio::Atom {
-            element: String::from("Fe"),
-            x: 4.0 * LATTICE_CONST,
-            y: 6.0 * LATTICE_CONST,
-            z: 8.0 * LATTICE_CONST,
-        });
-        atoms.push(xyzio::Atom {
-            element: String::from("Fe"),
-            x: 2.0 * LATTICE_CONST,
-            y: 1.0 * LATTICE_CONST,
-            z: 9.0 * LATTICE_CONST,
-        });
-        // get lattice size in each dimension
-        let sizes = auto_get_box_size(&atoms);
-        assert_eq!(sizes.0, 2 * 2);
-        assert_eq!(sizes.1, 5);
-        assert_eq!(sizes.2, 1);
-    }
-}
 
 #[cfg(test)]
 mod voronoy_tests {
@@ -325,20 +155,5 @@ mod voronoy_tests {
     fn test_voronoy() {
         assert_eq!(voronoy(1.377608, 1.501391, 1.471441), (1, 0, 0));
         assert_eq!(voronoy(2.772588, 0.056315, -0.044443), (2, 0, 0));
-    }
-}
-
-#[cfg(test)]
-mod cube_root_tests {
-    use super::*;
-
-    #[test]
-    fn test_cube_root() {
-        assert_eq!(cube_root(1), 1);
-        assert_eq!(cube_root(8), 2);
-        assert_eq!(cube_root(27), 3);
-        assert_eq!(cube_root(64), 4);
-        assert_eq!(cube_root(65), 0);
-        assert_eq!(cube_root(125), 5);
     }
 }
